move document, label, multilabel to haystack/schema.py
