+++ haystack/indexing/utils.py
@@ -11,7 +11,7 @@ from farm.data_handler.utils import http_get
 
 from haystack.indexing.file_converters.pdf import PDFToTextConverter
 from haystack.indexing.file_converters.tika import TikaConverter
-from haystack.documentstore.base import Document, Label
+from haystack import Document, Label
 
 logger = logging.getLogger(__name__)
