+++ test/test_dummy_retriever.py
@@ -1,4 +1,4 @@
-from haystack.documentstore.base import Document
+from haystack import Document
 import pytest
