+++ test/test_reader.py
@@ -1,6 +1,6 @@
 import math
 
-from haystack.documentstore.base import Document
+from haystack import Document
 from haystack.reader.base import BaseReader
 from haystack.reader.farm import FARMReader
 from haystack.reader.transformers import TransformersReader
